Update README.md
Dapp project from Ivan on Tech, now Moralis. I started project in early 2021. It took me around 3 months to complete but I went on to study React, which took me another three months. I was going to turn it into a React project but I am running out time...maybe later in the future. This project worked fine for me as a Ganache deployment but it is not minting on the Ropsten testnet....hopfully I can fix it soon.